Rename updateProperties.ts
